% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrevFuntions.R
\name{ac}
\alias{ac}
\title{Abbreviation for \code{as.character}}
\usage{
ac(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{Further arguments passed to \code{as.character}.}
}
\description{
Abbreviation for \code{as.character}
}

