% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfoFilename.R
\name{addInfoFilename}
\alias{addInfoFilename}
\title{Add extra info to file}
\usage{
addInfoFilename(albumFolder, extraInfo, musicExtension = ".mp3")
}
\arguments{
\item{albumFolder}{Folder where the files are located.}

\item{extraInfo}{String with the extra info that will be located at the end of the file.}

\item{musicExtension}{Extension of the files that will be matched for the substitution.}
}
\value{
\code{NULL}. This function just makes changes on filen names.
}
\description{
Add extra info to file
}
