% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{addTextBox}
\alias{addTextBox}
\title{Add a box with a text inside}
\usage{
addTextBox(xLimits, yLimits, text, border = NULL, col = "white",
  lty = par("lty"), ...)
}
\arguments{
\item{xLimits}{x-Axis limits for box.}

\item{yLimits}{y-Axis limits for box.}

\item{text}{A character or \link{expression} vector specifying the text to be written.}

\item{border}{the color to draw the border. The default, NULL, means to use par("fg"). Use border = NA to omit borders. See Details.}

\item{col}{The color for filling the box polygon.}

\item{lty}{The line type to be used, as in par.}

\item{...}{Extra arguments passed from \link{text} function.}
}
\description{
Add a box with a text inside
}
\details{
For extra details about \code{border}, \code{col} and \code{lty}, check the description of \link{polygon}.
}
\examples{
plot(1, 1, pch = NA)
addTextBox(xLimits = c(0.9, 1.1), yLimits = c(0.9, 1.1), text = "Hello World!", col = "indianred1", font = 2)
}
