% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{AIPData_new}
\alias{AIPData_new}
\title{\code{data.frame} with new (corrected) AIP information.}
\format{A \code{data.frame} with}
\usage{
AIPData_new
}
\description{
\code{data.frame} with 7 columns:
}
\references{
Instituto del Mar del Peru.
}

