% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ruisu.R
\name{AIPData_old}
\alias{AIPData_old}
\title{\code{data.frame} with old (original) AIP information.}
\format{
A \code{data.frame} with
}
\usage{
AIPData_old
}
\description{
\code{data.frame} with 7 columns:
}
\references{
Instituto del Mar del Peru.
}
