% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ruisu.R
\name{AIPShapefile_new}
\alias{AIPShapefile_new}
\title{\code{shapefile} with new (corrected) AIP information.}
\format{
A \code{SpatialPolygonsDataFrame}
}
\usage{
AIPShapefile_new
}
\description{
\code{SpatialPolygonsDataFrame} object.
}
\references{
Instituto del Mar del Peru.
}
