% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{AIPShapefile_old}
\alias{AIPShapefile_old}
\title{\code{shapefile} with old (original) AIP information.}
\format{A \code{SpatialPolygonsDataFrame}}
\usage{
AIPShapefile_old
}
\description{
\code{SpatialPolygonsDataFrame} object.
}
\references{
Instituto del Mar del Peru.
}
