% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{an}
\alias{an}
\title{Abbreviation for \code{as.numeric}}
\usage{
an(x, ...)
}
\arguments{
\item{x}{Object to be coerced or tested.}

\item{...}{Further arguments passed to \code{as.numeric}.}
}
\description{
Abbreviation for \code{as.numeric}
}
\examples{
as.numeric("02.33")
an("02.33")
}

