% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{anc}
\alias{anc}
\title{Abbreviation for \code{as.numeric(as.character(x))}}
\usage{
anc(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{Further arguments passed to \code{as.character}.}
}
\description{
Abbreviation for \code{as.numeric(as.character(x))}
}
\examples{
exampleVector <- runif(n = 20, min = 0, max = 100)
anc(cut(x = exampleVector, breaks = seq(0, 100, 20), labels = 1:5))
}
