% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPercentage.R
\name{barPercentage}
\alias{barPercentage}
\title{Pretty customizable barplot}
\usage{
barPercentage(x, cols = tim.colors(length(x)), lwd = 2, digits = 1,
  val_perc = c(1, 2), titleLab = NA, textPos = 2,
  sameColLines = TRUE, tagPosDelay = 0)
}
\arguments{
\item{x}{Numeric vector for plotting. See Details.}

\item{cols}{Color palette used for plotting.}

\item{lwd}{Line width.}

\item{digits}{Number of digits for rounding.}

\item{val_perc}{Whether to show values and/or percentage in labels.}

\item{titleLab}{Label for title.}

\item{textPos}{A position specifier for the text.}

\item{sameColLines}{\code{logical} Do you prefer that the color of lines and
the bar sections will be the same?}

\item{tagPosDelay}{Relative origin of guide lines expresed in a single number
from -1 to 1.}
}
\value{
Nothing but the pretty barplot.
}
\description{
Pretty customizable barplot
}
\details{
\code{x} must be numeric and the names of values will be used for
legend labels. If the vector does not have names, a temporal names will be
created on the form of 'Value_#'.
}
\examples{
barPercentage(x = c(55, 14, 10, 2))
}
