% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{bathymetry}
\alias{bathymetry}
\title{\code{data.frame} with topographic information.}
\format{
A \code{data.frame} with topographic information (lon, lat, altitude).
}
\usage{
bathymetry
}
\description{
A table with information of longitude, latitude and topography (high and depth) in meters.
}
