% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brokenBarplot.R
\name{brokenBarplot}
\alias{brokenBarplot}
\title{Make a barplot with a broken Y axis}
\usage{
brokenBarplot(
  height,
  ylim1,
  yAxis1 = list(),
  ylim2,
  yAxis2 = list(),
  names.arg = NULL,
  ...
)
}
\arguments{
\item{height}{Typical \code{height} parameter from \link{barplot}.}

\item{ylim1}{Limits Y axis of the top plot.}

\item{yAxis1}{Arguments passed to \link{axis} (Y axis) for the top plot.}

\item{ylim2}{Limits Y axis of the bottom plot.}

\item{yAxis2}{Arguments passed to \link{axis} (Y axis) for the bottom plot.}

\item{names.arg}{Typical \code{names.arg} parameter from \link{barplot}.}

\item{...}{Extra arguments passed to \link{barplot}.}
}
\description{
Make a barplot with a broken Y axis
}
\examples{
n <- 10
myData <- data.frame(x = letters[1:n], y = runif(n = n, min = 0, max = 50))
index <- c(3, 4, 8)
myData$y[index] <- runif(n = 3, min = 500, max = 1000)

brokenBarplot(height = myData$y, names.arg = myData$x,
              ylim1 = c(0, 50),
              yAxis1 = list(at = seq(0, 50, 10), las = 2),
              ylim2 = c(500, 1000),
              yAxis2 = list(at = seq(500, 1000, 100), las = 2))
}
