% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWaterMass.R
\name{calculateWaterMass}
\alias{calculateWaterMass}
\title{Title Function to identify water mass from different sources.}
\usage{
calculateWaterMass(data, method = "oliveros", sst = "sst", sss = "sss",
  lon = "lon", lat = "lat", month = "month", depth = "depth",
  dc = "dc", asFactors = TRUE)
}
\arguments{
\item{data}{\code{data.frame} including SST, SSS, month, depth, longitude and latitude information (View Details).}

\item{method}{Select method for water mass estimation: Oliveros (2020), Swartzman (2008) and Zuta (1978).}

\item{sst}{Name of column which contains Sea Surface Temperature info.}

\item{sss}{Name of column which contains Sea Surface Salinity info.}

\item{lon}{Name of column which contains Longitude info.}

\item{lat}{Name of column which contains Latitude info.}

\item{month}{Name of column which contains Month info.}

\item{depth}{Name of column which contains Depth info.}

\item{dc}{Name of column which contains Distance to Coast info.}

\item{asFactors}{\code{logical}. If \code{TRUE}, output will returned as a \code{factor} object, otherwise as
a \code{character} vector.}
}
\description{
Title Function to identify water mass from different sources.
}
\details{
Variables in \code{data} must be as next: SST (° S), SSS (PSU), depth (m), longitude (° W) and
latitude (° E).
}

