% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroidAssigner.R
\name{centroidAssigner}
\alias{centroidAssigner}
\title{Returns centroid values from grid codes.}
\usage{
centroidAssigner(code, what = "isoparalitoral", ...)
}
\arguments{
\item{code}{Vector with grid codes.}

\item{what}{\code{character} indicating whether AIP or Marsden squares are going to be used as reference.}

\item{...}{Extra arguments passed to selected method.}
}
\description{
This function takes a vector of grid codes and returns centroids (center of mass) in lon-lat values.
}
\details{
For \code{what = "isoparalitoral"} (default), it allows to use argument \code{old}, for specifying whether
to use old AIP shape (\code{AIPShapefile_old}) o the new (\code{AIPShapefile_new}).
}
\examples{
isopCodes <- c(1050, 4043, 17073, 27103)
centroidAssigner(code = isopCodes, what = "isoparalitoral")

marsdenCodes <- c("A6", "B8", "c12")
centroidAssigner(code = marsdenCodes, what = "marsden")
}
