% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{coastlineBuffer_m}
\alias{coastlineBuffer_m}
\title{\code{SpatialPolygonsDataFrame} object with a buffer in meters of South America coastline.}
\format{\code{SpatialPolygonsDataFrame} object with a buffer in meters of South America coastline.}
\usage{
coastlineBuffer_m
}
\description{
\code{SpatialPolygonsDataFrame} object with a buffer in meters of South America coastline.
}
