% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{coastlineBuffer_nm}
\alias{coastlineBuffer_nm}
\title{\code{SpatialPolygonsDataFrame} object with a buffer in nautical miles of South America coastline.}
\format{\code{SpatialPolygonsDataFrame} object with a buffer in nautical miles of South America coastline.}
\usage{
coastlineBuffer_nm
}
\description{
\code{SpatialPolygonsDataFrame} object with a buffer in nautical miles of South America coastline.
}
