% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterZOO.R
\name{counterZOO}
\alias{counterZOO}
\title{counterZOO}
\usage{
counterZOO(spList = NULL, filepath = "conteo.csv", reset = TRUE)
}
\arguments{
\item{spList}{Dirección de archivo que se desea utilizar para continuar el conteo. Si es
\code{NULL}, se iniciará un conteo nuevo, solicitando el ingreso de una especie. Este archivo
corresponderá a una tabla csv con las columnas "Especie", "Caracter", "Conteo", en ese orden.}

\item{filepath}{Nombre del archivo en donde se guardarán los resultados.}

\item{reset}{¿Desea continuar con el conteo anterior o colocar todos los valores en cero?}
}
\description{
Función simuladora de contador de organismos manual.
}
\details{
Esta función utiliza los siguientes comandos:
\itemize{
\item{\strong{new}{Para definir una especie nueva en el conteo.}}
\item{\strong{correct}{Para corregir o realizar una operación sobre alguno de los valores
de conteo. E.g. si se desea restar tres dígitos, se tendrá que digitar -3; si se desea sumar 2, +2.}}
\item{\strong{end}{Finaliza el conteo: Muestra y guarda la tabla de conteo.}}
\item{{\strong{help}}{Muestra la ayuda de comandos en consola.}}
}
}
\examples{
\dontrun{
counterZOO(spList = "myFolder/myTable.csv")
}
}

