% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitSum.R
\name{digitsum}
\alias{digitsum}
\title{Devuelve la suma de los dígitos de un número.}
\usage{
digitsum(x, recursive = FALSE)
}
\arguments{
\item{x}{Número o vector de números enteros o fraccionarios.}

\item{recursive}{¿Desea que la suma se realice hasta que el resultado conste de un dígito?}
}
\description{
Función (recreativa) para separar los dígitos de un número y obtener
la suma de ellos. Si \code{recursive = TRUE}, la suma se realizará hasta que el
resultado final conste de una sola cifra.
}
\examples{
digitsum(1516)

digitsum(1516, recursive = TRUE)
}

