% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{envirScenarios}
\alias{envirScenarios}
\title{\code{data.frame} with environmental scenarios given by IGP}
\format{A \code{data.frame} with environmental scenarios given by IGP (year_start, month_start,
year_end, month_end, magnitude, what).}
\usage{
envirScenarios
}
\description{
A table with information of El Nino/La Nina events classification from 1951 to 2016.
}
