% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAIPInfo.R
\name{getAIPInfo}
\alias{getAIPInfo}
\title{Extract info from AIP codes}
\usage{
getAIPInfo(aipVector)
}
\arguments{
\item{aipVector}{Vector of AIP.See Details}
}
\value{
A \code{data.frame} with variables dc, lat and upDown.
}
\description{
Extract info from AIP codes
}
\details{
AIP codes must be written in format DDLLPP, where DD is value of Distance to coast (mn/10),
LL are values of Latitude (as integer) and PP is position (up or down).
}
\examples{
getAIPInfo(c(30073, 1020, 2010))
}
