% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{getCoordsAxes}
\alias{getCoordsAxes}
\title{getCoordsAxes}
\usage{
getCoordsAxes(coord, what)
}
\arguments{
\item{coord}{Numeric vector of coordinates.}

\item{what}{Indicate \code{coord} belongs to longitude (\code{what = "lon"}) or
latitude (\code{what = "lat"}) values.}
}
\value{
A \code{character} vector, ready to put as coords labels.
}
\description{
getCoordsAxes
}
\examples{
myLongitudes <- seq(-20, -15, 0.5)
getCoordsAxes(coord = myLongitudes, what = "lat")
}

