% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnvirData.R
\name{getEnvirData}
\alias{getEnvirData}
\title{Get environmental info from lon-lat-time data frame}
\usage{
getEnvirData(
  x,
  environmentalDir,
  dimVars_x = list(lon = "lon", lat = "lat", date = "date", date_format = "\%F"),
  dimVars_nc = list(lon = "longitude", lat = "latitude", date = "time", date_origin =
    "1970-01-01T00:00:00Z"),
  regridSize = NULL,
  groupBy = "day",
  varName = NULL,
  justVarValues = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A \code{data.frame} with spatial-time information.}

\item{environmentalDir}{A \code{character} string for nc files.}

\item{dimVars_x}{A \code{list} with info about names (or position) for lon-lat-time columns on \code{x}. See Details.}

\item{dimVars_nc}{A \code{list} with info about names for lon-lat-time on nc files.}

\item{regridSize}{\code{NULL} by default. But if not, it indicates the size of a new grid, which is going to be applied
to data.}

\item{groupBy}{\code{character} string indicating whether to use data by 'month' or 'day' (default). See Details.}

\item{varName}{\code{character} strings with the name of the value for nc data. See Details.}

\item{justVarValues}{If \code{TRUE} (default) the function will return just a vector with the environmental values,
otherwise it will be the original \code{data.frame} plus the output as a extra column.}

\item{quiet}{If TRUE, suppress status messages as a progress bar.}
}
\value{
Depending on \code{justVarValues}, it could be a \code{numeric} vector or a \code{data.frame}.
}
\description{
Get environmental info from lon-lat-time data frame
}
\details{
\code{dimVars_x} and \code{dimVars_nc} requires user specify \code{date_format} and \code{date_origin} for
original data frame and nc data, respectively. For details about valid values for this levels, please check
\code{\link{strptime}} and \code{\link{as.POSIXct}} help.

This funtion has been developed for using nc files downloaded from ERDDAP server. Please check that your nc files
before to use.

Environmental data must be storaged as nc files which each file will contain daily info for a certain month, so if the
user has monthly data, it must be (externally) splited by day (repeating the monthly values) and save in monthly files.
If \code{groupBy="month"}, the function will average the matrices

If \code{varName} is not specified, the function will take the only available variable (if nc file has only one),
otherwise the function will stop and show the available variables for the nc. For this last option, function will
use a random file in order to get and show available variables, so it highly recomended not to put more than one
kind of nc files into the same directory.
}
