% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHarbor.R
\name{getHarbor}
\alias{getHarbor}
\title{Get harbor information  from name}
\usage{
getHarbor(myHarbor)
}
\arguments{
\item{myHarbor}{\code{character}. Text with the name of harbor.}
}
\description{
Get harbor information  from name
}
\examples{
getHarbor(myHarbor = "T. de mora")
}

