% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{getOverlay}
\alias{getOverlay}
\title{Get an overlay polygon from two set of points}
\usage{
getOverlay(points1, points2, fillBase = NULL)
}
\arguments{
\item{points1}{Set of points 1 (\code{data.frame} or \code{list}).}

\item{points2}{Set of points 2 (\code{data.frame} or \code{list}).}

\item{fillBase}{Indicates the way for closing the polygon. If \code{NULL} (default), the polygons will be closed
joining the first and the last point. For extra ways, see details.}
}
\value{
An \code{SpatialPolygons} object.
}
\details{
\code{points1} and \code{points2} must be a \code{list} or \code{data.frame} with 'x' and 'y' levels as
coordinates.
\code{fillBase} argument can be used for closing the polygons. If \code{NULL}, first and last points will
be joined. another value may be a function (\code{min}, \code{max}, \code{mean} or \code{median}) or a single
number (e.g. zero).
}
\examples{
n <- 30
points1 <- list(x = seq(n), y = runif(n = n, min = 20, max = 40))
points2 <- list(x = seq(n), y = runif(n = n, min = 5, max = 30))

plot(points1, type = "l", ylim = c(0, 40))
lines(points2)

overShape <- getOverlay(points1 = points1, points2 = points2, fillBase = 0)
}
