% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportion.R
\name{getProportion}
\alias{getProportion}
\title{Función para obtener proporciones a partir de una tabla}
\usage{
getProportion(table, group = 3)
}
\arguments{
\item{table}{Tabla en modo \code{data.frame} o \code{matrix}. Todos los valores deben
ser numéricos.}

\item{group}{Desea que las proporciones se tomen por fila (\code{group = 1}), columna
(\code{group = 2}) o la tabla completa (\code{group = 3}, valor por defecto).}
}
\value{
Una objeto (clase \code{matrix}) de las mismas dimensiones que el original con
las proporciones solicitadas
}
\description{
Toma una tabla y devuelve los valores proporcionales ya sea por fila,
columna o la tabla completa.
}
\examples{
myTable <- matrix(data = runif(n = 15, min = 0, max = 30), ncol = 3)
getProportion(table = myTable)
}

