% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{getSatellitalMaps}
\alias{getSatellitalMaps}
\title{Get satellital information for oceanographic variables}
\usage{
getSatellitalMaps(initialDate = NULL, finalDate = NULL, timeRes = "month",
  what = "SST", atDepth = NULL, dateList = NULL, lonRange = c(-85, -70),
  latRange = c(-20, -2), outputFormat = "png", outputDir = ".",
  showURL = FALSE)
}
\arguments{
\item{initialDate}{Initial date.}

\item{finalDate}{Final date.}

\item{timeRes}{Time resolution for maps: Monthly ('month') o daily ('day').}

\item{what}{What variable do you want to download? See details.}

\item{atDepth}{Depth (usually meters) of layer to downloading.}

\item{dateList}{List with dates for downloading..}

\item{lonRange}{Range of longitude. If \code{NULL}, it will take values from 85 W to 90 W (Peru).}

\item{latRange}{Range of latitude. If \code{NULL}, it will take values from 2 S to 20 S (Peru).}

\item{outputFormat}{What extension would you like to use for download? (nc, csv, png).}

\item{outputDir}{Folder for downloading.}

\item{showURL}{Do you want to show the URL each time?}
}
\description{
This function uses date and space range and get maps for Sea Surface Temperature (SST),
Sea Surface Salinity (SSS), Chlorophyll-a (Chl), Topography (topo), Sea level (sealevel) and SODA
estimations for Temperature (sst_soda) and Salinity (sss_soda) by depth (\code{atDepth} parameter).
}
\details{
Function will use data bases of ERDDAP website
\link{https://coastwatch.pfeg.noaa.gov/erddap/info/index.html?page=1&itemsPerPage=1000}.

This function will use the dates values to choose from what source the values will be taken.
}
\examples{
getSatellitalMaps(initialDate = "2014-3-26", finalDate = "2014-4-17", what = "sst", timeRes = "day",
lonRange = c(-90, -85), latRange = c(-5, 0), outputFormat = "png",
outputDir = "../")
}

