% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{getTopography}
\alias{getTopography}
\title{Get topogaphy values from an XY table}
\usage{
getTopography(allData, colLon = "lon", colLat = "lat", units = "m",
  useInternet = TRUE)
}
\arguments{
\item{allData}{\code{matrix} or \code{data.frame} where the points are.}

\item{colLon}{Name of column which contains Longitude info.}

\item{colLat}{Name of column which contains Latitude info.}

\item{units}{Define the unit for outputs: nm (nautical miles), kilometers (km), m (meters, default).}

\item{useInternet}{\code{logical} indicating if the function will use an online source for some points out of limits. See Details.}
}
\value{
A \code{numeric} vector with depth values.
}
\description{
Get topogaphy values from an XY table
}
\details{
The function will use a fixed data base (\code{bathymetry}) which only has information for topography from 100 to 70 W
and from 20 to 0 S. If there were points outside these limits, the function could use (depending on \code{useInternet} argument)
an external source (\link{https://coastwatch.pfeg.noaa.gov/erddap/griddap/etopo180.graph}) for downloading information about topography
for completing the output depths.
}
\examples{
exampleData <- data.frame(lon = c(-80.757, -80.284, -79.511, -78.738, -79.511),
                          lat = c(-5.794, -7.149, -8.422, -9.983, -11.011),
                          stringsAsFactors = FALSE)

exampleData$depth <- getTopography(allData = exampleData)
}
