% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermass_functions.R
\name{getWatermasses}
\alias{getWatermasses}
\title{Title}
\usage{
getWatermasses(x, criteria = "swartzman2008", newCriteria = NULL, ...)
}
\arguments{
\item{x}{A \code{data.frame} containing the variables required by the indicated \code{criteria}.}

\item{criteria}{A \code{chracter} indicating the method used for the definition.}

\item{newCriteria}{A \code{list} with information for a new watermass definition. See Details.}

\item{...}{Extra arguments (at moment, with no sense).}
}
\value{
A \code{factor} vector with the definitions.
}
\description{
Title
}
\details{
The \code{newCriteria} argument allows the users to define their own criteria for classify the
water masses. To do that, \code{newCriteria} must be a list with values of ranges and definitions, just
like \code{swartzman2008_ranges} and \code{swartzman2008_definitions} objects.
}
