% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ruisu.R
\name{harborData}
\alias{harborData}
\title{\code{data.frame} with information of the main Peruvian harbors}
\format{
A \code{data.frame} with information of Peruvian harbors: name, pattern, lon, lat, area, importance. See Details
}
\usage{
harborData
}
\description{
Some detailed information of the main Peruvian harbors.
}
\details{
\code{pattern} information is used by \code{getHarbor} funtion for making a match between given strings and
the harbor names of \code{harbordata}.
\code{importance} column separates the principal harbors (1 value) from the secondary ones (0 values).
}
\references{
Instituto del Mar del Peru.
}
