% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{isNearCoast}
\alias{isNearCoast}
\title{Determine the proximity to coast of points}
\usage{
isNearCoast(dataPoints, colLon = "lon", colLat = "lat", units = "m",
  distance = 20)
}
\arguments{
\item{dataPoints}{\code{data.frame} which has Longitude and Latitude information.}

\item{colLon}{Name of column which contains Longitude info.}

\item{colLat}{Name of column which contains Latitude info.}

\item{units}{Which units do you want to use for measuring the proximity to coast.}

\item{distance}{Maximum distance (in degrees) to make the searching.}
}
\value{
A \code{logical} vector indicating whether the coordinates belong to the area between
the coast line and a buffer of the selected distance.
}
\description{
Determine the proximity to coast of points
}
\examples{
isNearCoast(c(-81.191, -5.211), distance = 20)
isNearCoast(c(-81.191, -5.211), distance = 10)
}
