% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{isopArea.assigner}
\alias{isopArea.assigner}
\title{Title Function to get AIP code from lon-lat information.}
\usage{
isopArea.assigner(dataPoints, colLon = "lon", colLat = "lat", old = TRUE)
}
\arguments{
\item{dataPoints}{\code{data.frame} which has Longitude and Latitude information.}

\item{colLon}{Name of column which contains Longitude info.}

\item{colLat}{Name of column which contains Latitude info.}

\item{old}{\code{logical}. Do you prefer to use old or new AIP data?}
}
\value{
A numeric vector indicating the AIP values for each coord.
}
\examples{
exampleCoords <- data.frame(lon = runif(n = 10, min = -80, max = -78),
                            lat = runif(n = 10, min = -14, max = -12))

isopArea.assigner(dataPoints = exampleCoords)
}
