% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{makeWindPlot}
\alias{makeWindPlot}
\title{Plot typical wind arrows from wind velocity data}
\usage{
makeWindPlot(uComponent, vComponent, maxLength = 1, densityfactor = 0.98,
  arrowCol = "black", add = TRUE, includeRaster = TRUE,
  col = tim.colors(1000), xInterval = NULL, yInterval = NULL, ...)
}
\arguments{
\item{uComponent}{Matrix with values of velocity in x component See Details.}

\item{vComponent}{Matrix with values of velocity in y component See Details.}

\item{maxLength}{For changing length of arrows (\code{numeric}).}

\item{densityfactor}{For changing the quantity of arrows on plot. Value from 0 to 1.}

\item{arrowCol}{Color of arrows.}

\item{add}{\code{logical}; if \code{TRUE}, add to current plot and \code{...} will not be considered.}

\item{col}{Color table to use for image.}

\item{xInterval}{If \code{add = TRUE}, it works for specifing the interval of marks in X axis.}

\item{yInterval}{If \code{add = TRUE}, it works for specifing the interval of marks in Y axis.}

\item{...}{Extra arguments passed to \code{image} function. It will only be used if \code{add = FALSE}.}
}
\value{
A plot of arrows indicating the intensity and direction of winds.
}
\details{
Both \code{uComponent} and \code{vComponent} must be list objects with levels x (vector of longitudes),
y (vector of latitudes) and z (matrix of values).
}
