% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ruisu.R
\name{month.abb_spanish}
\alias{month.abb_spanish}
\title{\code{charcater} vector with the three-letter abbreviations for names of months in Spanish.}
\format{
A \code{charcater} vector with the three-letter abbreviations for names of months in Spanish.
}
\usage{
month.abb_spanish
}
\description{
Similar to \code{month.abb} but in Spanish.
}
