% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{month.name_spanish}
\alias{month.name_spanish}
\title{\code{charcater} vector with names of months in Spanish.}
\format{
A \code{charcater} vector with names of months in Spanish.
}
\usage{
month.name_spanish
}
\description{
Similar to \code{month.name} but in Spanish.
}
