% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverage.R
\name{movingAverage}
\alias{movingAverage}
\title{Calculate the moving average for a numeric vector}
\usage{
movingAverage(x, n = 3, circular = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{n}{Size for grouping.}

\item{circular}{If \code{TRUE}, wrap the filter around the ends of the series,
otherwise assume external values are missing (\code{NA}).}

\item{...}{Extra arguments passed to \code{\link{filter}} function.}
}
\value{
A numeric vector with the same length of the original vector.
}
\description{
Calculate the moving average for a numeric vector
}
\examples{
exampleVector <- runif(n = 20, min = 0, max = 100)
movingAverage(x = exampleVector)
}
