% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newYear.R
\name{newYear}
\alias{newYear}
\title{Curious way to show a message.}
\usage{
newYear(message, delay = 4, nroBombs = 100, dispersion = 10, cex.text = 3)
}
\arguments{
\item{message}{What message do you want to show?}

\item{delay}{Specify the time delay between letters.}

\item{nroBombs}{Number of lights (points) that will be shown.}

\item{dispersion}{Dispersion of points.}

\item{cex.text}{Size of text.}
}
\description{
Curious way to show a message.
}
\examples{
newYear("Happy new year_2017!")
}
