% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{notebanks_ATM}
\alias{notebanks_ATM}
\title{ATM banknotes' optimizer algorithm}
\usage{
notebanks_ATM(amount, d_Banknotes, n_Banknotes, d_symbol = "US$")
}
\arguments{
\item{amount}{The amount to optimize.}

\item{d_Banknotes}{Denomination of banknotes.}

\item{n_Banknotes}{Quantity of available banknotes.}

\item{d_symbol}{Symbol of banknote currency.}
}
\value{
A vector with the quantity of banknotes.
}
\description{
ATM banknotes' optimizer algorithm
}
\details{
For now, the algorithm compares all the possible solutions, so it is a brute-force search.
}
\examples{
amount <- 70

d_Banknotes <- c(20, 50, 100, 200)
n_Banknotes <- rep(10, length(d_Banknotes))

notebanks_ATM(amount = amount, d_Banknotes = d_Banknotes,
              n_Banknotes = n_Banknotes, d_symbol = "S/")
}
