% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruisu.R
\docType{package}
\name{oXim-package}
\alias{oXim}
\alias{oXim-package}
\title{Oxycline Index from Matrix Echograms}
\description{
Pack oriented to extract oxycline depth from echogram matrix using Sobel-based algorithms.
}
\author{
Wencheng Lau-Medrano, \email{llau@imarpe.gob.pe}
}
\references{
oXim: Oxycline Index from Matrix Echograms (RJournal)
}
\keyword{depth,}
\keyword{echograms,}
\keyword{image-filtering}
\keyword{oxycline,}

