% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p2pDistance}
\alias{p2pDistance}
\title{Función para obtener distancia entre dos puntos.}
\usage{
p2pDistance(dataFrom, dataTo, greatCircle = FALSE)
}
\arguments{
\item{dataFrom}{\code{data.frame} with lon-lat values.}

\item{dataTo}{Reference lon-lat \code{data.frame}.}

\item{greatCircle}{\code{logical}; si es \code{FALSE}, se devolverá la distancia euclidiana: si es
\code{TRUE} devolverá la DIstancia de Gran Círculo (WGS84).}
}

