% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralelMaps.R
\name{paralelMaps}
\alias{paralelMaps}
\title{Function for pltting Parallel maps}
\usage{
paralelMaps(file, colCategories, allCategories = NULL,
  categoriesNames = NULL, colLon = "lon", colLat = "lat",
  xlim = NULL, ylim = NULL, yInterval = NULL, xDelay = NULL,
  newPlot = FALSE, seaCol = "lightblue1",
  landCol = "lightgoldenrod1", showBoxText = TRUE,
  textBoxLimits = NULL, textBox = NULL, textBox_line = -2,
  textBox_cex = 1.2, textBox_font = 2, textBoxXCF = 0,
  catNames_col = "black", catNames_cex = 1, catNames_font = 2,
  addHarbors = TRUE)
}
\arguments{
\item{file}{Indicates the table from the main length distribution will be drawn. See details.}

\item{colCategories}{Name or number of columns which will be used for make categorization.}

\item{allCategories}{Vector for categories (optional). It must contain all values on 'colCategories'.}

\item{categoriesNames}{\code{character} vector indicating the names (subtitle) showed for each sub map.}

\item{colLon}{Name or position of column for longitude. As default, it will be \code{lon}.}

\item{colLat}{Name or position of column for latitude As default, it will be \code{lat}.}

\item{xlim}{Limits of X axis.}

\item{ylim}{Limits of Y axis.}

\item{yInterval}{Number of intervals for y axis labels.}

\item{xDelay}{Argument which controls separation between maps.}

\item{newPlot}{If \code{TRUE}, the plot will be opened in a new window (using \code{x11} command).}

\item{seaCol}{Color for sea background.}

\item{landCol}{Color for land.}

\item{showBoxText}{If \code{TRUE} (default) it allows to show the names of categories for each sub map.}

\item{textBoxLimits}{A \code{list} (two vector: x and y) with values for limiting text box.}

\item{textBox}{String with text included inside the text box.}

\item{textBox_line}{\code{line} (see \code{\link{mtext}}) parameter for text box.}

\item{textBox_cex}{\code{cex} (see \code{\link{mtext}}) parameter for text box.}

\item{textBox_font}{\code{font} (see \code{\link{mtext}}) parameter for text box.}

\item{textBoxXCF}{Correction Factor on X for text box.}

\item{catNames_col}{\code{col} (see \code{\link{mtext}}) parameter for Categories' text.}

\item{catNames_cex}{\code{cex} (see \code{\link{mtext}}) parameter for Categories' text.}

\item{catNames_font}{\code{font} (see \code{\link{mtext}}) parameter for Categories' text.}

\item{addHarbors}{\code{logical}. Add harbor text/points over the coastline?}
}
\value{
A plot with parallel maps.
}
\description{
Function for pltting Parallel maps
}
