% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelMaps.R
\name{parallelMaps}
\alias{parallelMaps}
\title{Make plots with parallel effect}
\usage{
parallelMaps(
  x,
  overlay = 0,
  direction = 2,
  labels.args = list(),
  map.args = list(),
  final.par = list(),
  saveFig = FALSE,
  png.args = list(),
  extra.args = list(res = 450, width = 2000),
  xlim = NULL,
  ylim = NULL,
  col = "black",
  pch = 1,
  cex = 1,
  interpolate = FALSE,
  show.axis.labels = TRUE,
  pos.axis.labels = ifelse(direction\%\%2 == 0, 2, 1),
  axis.labels.args = list(),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} with coordinates and group column.}

\item{overlay}{\code{numeric} value indicating the level of overlay between
subplots (\eqn{[0; 0.5>}).}

\item{direction}{\code{numeric} On which side of the plot (1 = bottom,
2 = left, 3 = top, 4 = right) the subplots will be placed?}

\item{labels.args}{Extra arguments for \link[graphics]{mtext} that is used to
specify the labels of groups. If \code{NULL}, labels will not be plotted.}

\item{map.args}{Extra arguments for \link[maps]{map} that is used to add map
shapes to subplots. If \code{NULL}, map lines will not be plotted.}

\item{final.par}{\code{list} with graphical settings applied to the final
plot. See the available options in \link[graphics]{par}.}

\item{saveFig}{\code{logical} Do you want to save final plot as a .png file?
If \code{saveFig = TRUE} (default), the plot will not be showed.}

\item{png.args}{\code{logical} If \code{saveFig = TRUE}, extra arguments
passed to \link[grDevices]{png}.}

\item{extra.args}{Extra arguments passed to some internal steps. See Details.}

\item{xlim, ylim}{\code{numeric} Range of subplots in X and Y axis.}

\item{col, pch, cex}{Features that will vary by groups.}

\item{interpolate}{a \code{logical} vector (or scalar) indicating whether to
apply linear interpolation to the image when drawing. Argument passed from
\link[graphics]{rasterImage}.}

\item{show.axis.labels}{\code{logical} that defines if the axis labels will
be plotted.}

\item{pos.axis.labels}{If \code{show.axis.labels = TRUE}, this arguments
indicates the position of the axis labels.}

\item{axis.labels.args}{If \code{show.axis.labels = TRUE}, extra arguments
passed to \link[graphics]{axis}.}

\item{quiet}{\code{logical} value to control informative messages.}

\item{...}{Extra arguments passed to \link{plot} that makes subplots.}
}
\value{
Depending on \code{saveFig} and \code{plot} values, this function
will return a plot
}
\description{
This function takes a \code{data.frame} of coordinates and uses a 'group'
column to make a plot placing the points in a parallel way.
}
\details{
\code{x} must be defined as a \code{data.frame} with at least 3 columns: lon,
lat and group. It is recommended that to define \strong{group} as a factor,
otherwise, it will be defined internaly as it using \link{as.factor}.

This function plots and saves internaly one figure per group. The quality of
this subplots can be modified by \code{extra.args} setting \code{res}
(default 200) and \code{width} in pixels (default 2500).
}
\examples{
data(coord_example)

parallelMaps(x = coord_example, direction = 4, overlay = 0.4,
             xlim = c(-82, -74), ylim = c(-16, -5),
             labels.args = list(adj = 0.5, side = 1),
             col = c("blue", "red"), pch = c(16, 17),
             map.args = list(database = "worldHires"))
}
