% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passwordGen.R
\name{passwordGen}
\alias{passwordGen}
\title{Password generator}
\usage{
passwordGen(
  len = 10,
  numbers = TRUE,
  special_chars = TRUE,
  upper_letters = TRUE
)
}
\arguments{
\item{len}{\code{numeric} Total length of the required password.}

\item{numbers, special_chars, upper_letters}{\code{logical} Do you want to
include some of this elements in your password? \code{TRUE} by default.}
}
\value{
A character vector of length 1.
}
\description{
A quite simple password generator.
}
\details{
If the user just specifies the lenght, \code{passwordGen} will return a
password composed with 20% of numbers, 20% of special characters, 30% of
lower-case and 30% of upper-case letters.

\code{passwordGen} will generate one password at time. If you want to
generate several passwords, you can use \code{sapply(rep(7, 10), passwordGen)}
to generate in that case 10 passwords of length 7.
}
\examples{
passwordGen(len = 10)
}
