% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIsoparalitoral.R
\name{plotIsoparalitoral}
\alias{plotIsoparalitoral}
\title{Title plotIsoparalitoral}
\usage{
plotIsoparalitoral(codeList = NULL, add = FALSE, old = TRUE,
  plotParams = list(xlim = c(-90, -70), ylim = c(-18, -2), axes = TRUE),
  mapParams = list(database = "world"), ...)
}
\arguments{
\item{codeList}{Vector conteniendo los códigos de AIP a plotear.}

\item{add}{¿Desea agregar las AIP a un gráfico presente (\code{TRUE}) o desea plotear un mapa
con líneas de costa (\code{FALSE})?}

\item{old}{¿Desea utilizar las AIP anteriores (\code{AIPShapefile_old}) o las nuevas (\code{AIPShapefile_new})?}

\item{plotParams}{Objeto \code{lista} conteniendo parámetros adicionales pasados a la función \code{plot}. Se
usará solo si \code{add = FALSE}.}

\item{mapParams}{Objeto \code{lista} conteniendo parámetros adicionales pasados a la función \code{map}. Se
usará solo si \code{add = FALSE}.}

\item{...}{Parámetros adicionales pasados a la función \code{\link{plot.SpatialPolygons}}.}
}
\description{
Title plotIsoparalitoral
}

