% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{prepareProjFolder}
\alias{prepareProjFolder}
\title{prepareProjFolder}
\usage{
prepareProjFolder(folder, type = 1)
}
\arguments{
\item{folder}{\code{character}. Path of the folder that will be created.}

\item{type}{Indicates the type of sub directories that will be created (see Details).}
}
\description{
prepareProjFolder
}
\examples{
\dontrun{
prepareProjFolder(folder = "../exampleFolder", type = 3)
}
}

