% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{progressBar}
\alias{progressBar}
\title{Make a progress bar}
\usage{
progressBar(i, n = NULL, iterator = NULL)
}
\arguments{
\item{i}{Step of the iteration.}

\item{n}{Total number of iteration}

\item{iterator}{If 'i' belongs to a vector of iterators, define it here.}
}
\value{
The function returns only messages (from \code{cat}).
}
\description{
Make a progress bar
}
\examples{
n = 1234
for(i in seq(n)){
  Sys.sleep(0.01)
  progressBar(i = i, n = n)
}
}
