% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{randomRectangles}
\alias{randomRectangles}
\title{Title Draw colorful squares}
\usage{
randomRectangles(nsquares = 500)
}
\arguments{
\item{nsquares}{How much squares do you desire? (500, as default).}
}
\description{
Title Draw colorful squares
}
\examples{
randomRectangles(nsquares = 120)
}

