% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruisu-package.R
\docType{package}
\name{ruisu-package}
\alias{ruisu}
\alias{ruisu-package}
\title{Tools for habitual tasks (and another fancy things) in data analysis}
\description{
Tools for common tasks as well as some fancy functions
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@outlook.com}
}
\keyword{miscellany}
