% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ruisu.R, R/sillyTest.R
\name{sillyTest}
\alias{sillyTest}
\title{\code{list} with a test inspired on Harry Potter books.}
\format{
A \code{list} with two levels: 1) The first, the questions as a \code{character} vector and
2) the second, the posible answers for each question.
}
\usage{
sillyTest

sillyTest(testName, questions, answers, correctAnswers)
}
\arguments{
\item{testName}{The name for the test}

\item{questions}{\code{character} vector indicating the questions.}

\item{answers}{\code{list} with the options for each question.}

\item{correctAnswers}{\code{numeric} vector indicating the index for correct answers.}
}
\value{
Nothing.
}
\description{
A \code{list} including Q&A inspired on Harry Potter books in order to clarify how
to organize inputs for \code{sillyTest} function.
}
\details{
Of course, arguments for questions, answers and correct answers must have the same length.
}
\examples{
\dontrun{
sillyTest(testName = "Harry Potter", questions = testExample$questions,
          answers = testExample$answers, correctAnswers = testExample$correctAnswers)
}
}
