% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{speciesInfo}
\alias{speciesInfo}
\title{\code{data.frame} with main information about some Peruvian fishes.}
\format{A \code{data.frame} with main information about some Peruvian fishes.}
\usage{
speciesInfo
}
\description{
A table with information of scientific name, type of measurement of length, typycal unit
for expresing this measurements, Minimum length, Maximum length, typical interval length, juvenile length.
}
