% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyTrack.R
\name{surveyTrack}
\alias{surveyTrack}
\title{Simulate points for a survey track}
\usage{
surveyTrack(breakPoints, dotDensity, dcLength, legLength)
}
\arguments{
\item{breakPoints}{Arguments for changing the position of change for track way. See Details.}

\item{dotDensity}{Density of points for tracks in dots-per-nm.}

\item{dcLength}{Distance from coast in nm (nautical miles) of tracks.}

\item{legLength}{Length of tracks in nm.}
}
\value{
A \code{data.frame} with 4 columns: n, lon, lat and index.
}
\description{
This function simulates the track points of a typical acoustic survey.
}
\details{
\code{breakpoints} must be a list of 3 levels:
\itemize{
\item \strong{points}, which indicates the latitude limits for track sections.
\item \strong{angle}, indicating the angle of tracks.
\item \strong{n_tracks}, how many tracks by section do you want?
}
}
\examples{
breakPoints <- list(points = c(-4, -6, -15.4, -18.3),
                    angle = c(180, 210, 240),
                    n_tracks = c(9, 39, 18))
dotDensity  <- 0.1
dcLength    <- 5
legLength   <- 100

output <- surveyTrack(breakPoints = breakPoints, dotDensity = dotDensity,
                      dcLength = dcLength, legLength = legLength)

plot(output$lon, output$lat, pch = 19, cex = 0.1)
}
