% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableAdd.R
\name{tableAdd}
\alias{tableAdd}
\title{Add info (total, mean, sd, etc.) to matrix or table}
\usage{
tableAdd(x, side = c(1, 4), FUN = sum, label = "Total", ...)
}
\arguments{
\item{x}{\code{data.frame} or \code{matrix} object. All values must be \code{numeric}.}

\item{side}{Indicates the side (1: bottom, 2: left, 3: top, 4: right) for adding info.}

\item{FUN}{a function (or name of a function) to be applied. The indicated function
must return \code{atomic} vectors.}

\item{label}{What name do you want to use for extra column/row?}

\item{...}{Extra arguments passed to \code{FUN.}}
}
\value{
Same class object of \code{x}.
}
\description{
This functions add values from a certain function to the beggining or
end of a table.
}
\examples{
tableAdd(x = mtcars)
}
