% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{testExample}
\alias{testExample}
\title{\code{list} with a test inspired on Harry Potter books.}
\format{
A \code{list} with two levels: 1) The first, the questions as a \code{character} vector and
2) the second, the posible answers for each question.
}
\usage{
testExample
}
\description{
A \code{list} including Q&A inspired on Harry Potter books in order to clarify how
to organize inputs for \code{sillyTest} function.
}
