% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waiting.R
\name{waiting}
\alias{waiting}
\title{Function that runs (waits) until a specified time has elapsed}
\usage{
waiting(time_ref, time_check = 60, msg = NULL, ...)
}
\arguments{
\item{time_ref}{Time limit after which the function will end. Internally,
\code{as.POSIXct} will be used to convert it to a date-time object.}

\item{time_check}{How often will \code{time_ref} be checked?}

\item{msg}{A message to be printed after the reference time has been reached.}

\item{...}{Extra arguments passed to \code{as.POSIXct}.}
}
\description{
Function that runs (waits) until a specified time has elapsed
}
\details{
Internally, \code{waiting} will use the local time (including the time zone).
It is important to bear this in mind before defining \code{time_ref}. For
example, if the time zone is UTC -05 and our PC is configured with this time
zone, the definition of \code{time_ref} should be something like
\code{“2025-09-12 09:00:00 -05”}. It is recommended to perform a quick test
\strong{before} implementing this function within a process.
}
\examples{
waiting(time_ref = Sys.time() + 5,
        time_check = 5,
        msg = "\nIt works!\n")
}
