% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{zigguratCreator}
\alias{zigguratCreator}
\title{Creates a zigurat matrix}
\usage{
zigguratCreator(n)
}
\arguments{
\item{n}{Number of floors of zigurat (size of matrix).}
}
\value{
A square matrix \eqn{n * n}.
}
\description{
Creates a zigurat matrix
}
\examples{
out <- zigguratCreator(n = 10)
image(out)
}
