% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{addCoordsAxes}
\alias{addCoordsAxes}
\title{Add labels for XY axis}
\usage{
addCoordsAxes(xParams = NULL, yParams = NULL, what = c("x", "y"))
}
\arguments{
\item{xParams}{Parameters for X axis.}

\item{yParams}{Parameters for Y axis.}

\item{what}{\code{character} vector indicating whether to show the labels for axis X, Y or both.}
}
\description{
Add labels for XY axis
}
\details{
\code{xParams} and \code{yParams} must contain axis information as a 3-length vector: \code{c(from, to, by)}.
For instance, to indicate from -100 S to -70 S by 5, the vector will be c(-100, -70, 5).
}
\examples{
par(mar = c(2, 3, 1, 1), xaxs = "i", yaxs = "i")

xlim <- c(-85, -70)
ylim <- c(-20, -2)

plot(1, 1, type = "n", axes = FALSE, xlab = NA, ylab = NA, xlim = xlim, ylim = ylim)

addCoordsAxes(yParams = c(ylim, 2))

box()
}
