% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCoordsAxes.R
\name{addCoordsAxes}
\alias{addCoordsAxes}
\title{Add labels for XY axis on a map}
\usage{
addCoordsAxes(
  xParams = NULL,
  yParams = NULL,
  where = c(1, 2),
  las = 1,
  labels = NULL,
  dms = "d",
  ...
)
}
\arguments{
\item{xParams}{Parameters for X axis.}

\item{yParams}{Parameters for Y axis.}

\item{where}{\code{character} vector indicating where to show the labels for axis: 1 (bottom), 2 (left),
3 (top) or 4 (right).}

\item{las}{\code{numeric} in {0,1,2,3}; the style of axis labels. See \link{par}.}

\item{labels}{\code{character} vector setting the labels for axis.}

\item{dms}{A \code{character} strings indicating what to show (d: degrees,
m: minutes, s: seconds).}

\item{...}{Extra arguments passed to \link{axis} function (e.g. \code{cex.axis}), which is used internaly for ploting.}
}
\description{
Add labels for XY axis on a map
}
\details{
\code{xParams} and \code{yParams} must contain axis information as a 3-length vector:
\code{c(from, to, by)}. For instance, to indicate from -100 S to -70 S by 5, the vector will be c(-100, -70, 5).

If \code{labels} is specified at \code{...}, the function will consider this values for both \code{xParams} and
\code{yParams}, so be carefull and use it separately (run the function twice for lon and lat).
}
\examples{
par(mar = c(2, 3, 1, 3), xaxs = "i", yaxs = "i")

xlim <- c(-85, -70)
ylim <- c(-20, -2)

plot(1, 1, type = "n", axes = FALSE, xlab = NA, ylab = NA, xlim = xlim, ylim = ylim)

addCoordsAxes(xParams = c(xlim, 5), yParams = c(ylim, 2), where = c(1, 2, 4))

box()
}
