% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{centroidAssigner}
\alias{centroidAssigner}
\title{Returns centroid values from Isoparalitoral-area codes.}
\usage{
centroidAssigner(isoCode, old = TRUE)
}
\arguments{
\item{isoCode}{Vector with AIP codes.}

\item{old}{\code{logical}. Specifying whether to use old AIP shape (\code{AIPShapefile_old}) o the new (\code{AIPShapefile_new}).}
}
\description{
This function takes a vector of AIP codes and returns centroids (center of mass) in lon-lat values.
}
\examples{
areaCodes <- c(1050, 4043, 17073, 27103)
centroidAssigner(isoCode = areaCodes)
}
