% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroidAssigner.R
\name{centroidAssigner}
\alias{centroidAssigner}
\title{Returns centroid values from Isoparalitoral-area codes.}
\usage{
centroidAssigner(isoCode, old = TRUE)
}
\arguments{
\item{isoCode}{Vector de códigos de AIP de los que se desea obtener  los centroides.}

\item{old}{\code{logical}. ¿Desea utilizar la base de datos antigua (\code{old = TRUE}) o nueva \code{old = FALSE}?}
}
\description{
Esta función toma un vector de códigos de AIP y retorna las coordenadas de los  centroides
(longitud, latitud). El parámetro \code{old} controla la base de AIP a utilizar: Si \code{old = TRUE},
se usará la base de AIP creada en llos 90'; de otro modo, se utilizará la creada en 2016.
}
\examples{
areaCodes <- c(1050, 4043, 17073, 27103)
centroidAssigner(isoCode = areaCodes)
}

