% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{coastline}
\alias{coastline}
\title{\code{data.frame} with coordinates of the eastern side of the American coast.}
\format{A \code{data.frame} with coordinates of the eastern side of the American coast: country, lon, lat}
\usage{
coastline
}
\description{
coordinates of the eastern side of the American coast.
}
\references{
Instituto del Mar del Peru.
}
