% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{data}
\name{coord_example}
\alias{coord_example}
\title{\code{data.frame} with some random coordinates for running \link[ruisu]{parallelMaps}.}
\format{
\code{data.frame} object with three columns: lon, lat and group.
}
\usage{
coord_example
}
\description{
\code{data.frame} with some random coordinates for running \link[ruisu]{parallelMaps}.
}
