% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterZOO.R
\name{counterZOO}
\alias{counterZOO}
\title{Laboratory Counter simulator}
\usage{
counterZOO(spList = NULL, filepath = "conteo.csv", reset = TRUE)
}
\arguments{
\item{spList}{(Optional) Path of file to re start counting. If \code{NULL}, new counting will start. This file
requires columns "Especie", "Caracter", "Conteo".}

\item{filepath}{File name where the output file will be saved.}

\item{reset}{Would you want to continue with the previous count (\code{FALSE}) or reset (put in zeros)?}
}
\description{
Laboratory Counter simulator
}
\details{
This function uses next commands:
\itemize{
\item{\strong{new}{Defines a new species that will be included in the count.}}
\item{\strong{correct}{Allows to make some basic operation in count values. E.g. +1, -2.}}
\item{\strong{end}{Ends the count and save the file.}}
\item{{\strong{help}}{Shows commands' help in console.}}
}
}
\examples{
\dontrun{
counterZOO(spList = "myFolder/myTable.csv")
}
}
