% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{digitsum}
\alias{digitsum}
\title{Returns sum of digit from number}
\usage{
digitsum(x, recursive = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{recursive}{Would you want to sum until output consist of one digit?}
}
\description{
Recreational function that split digit from number and gets sum of them. If \code{recursive = TRUE}
this operation will be done until output consist of one digit.
}
\examples{
digitsum(1516)

digitsum(1516, recursive = TRUE)
}

