% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoordsAxes.R
\name{getCoordsAxes}
\alias{getCoordsAxes}
\title{Easy way to get the labels of coordinates from a vector.}
\usage{
getCoordsAxes(coord, what)
}
\arguments{
\item{coord}{Numeric vector of coordinates.}

\item{what}{Indicate \code{coord} belongs to longitude (\code{what = "lon"}) or
latitude (\code{what = "lat"}) values.}
}
\value{
A \code{character} vector, ready to put as coords labels.
}
\description{
Easy way to get the labels of coordinates from a vector.
}
\examples{
myLongitudes <- seq(-20, -15, 0.5)
getCoordsAxes(coord = myLongitudes, what = "lat")
}
