% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{getProportion}
\alias{getProportion}
\title{Get proportions from table}
\usage{
getProportion(table, group = 3)
}
\arguments{
\item{table}{\code{data.frame} or \code{matrix} object. All values must be numeric.}

\item{group}{Specify how the proportions will be extracted. 1 indicates rows, 2 indicates
columns and 3 indicates rows and columns (total).}
}
\description{
Takes a table and gets propotions by row, column or total.
}
\examples{
myTable <- matrix(data = runif(n = 15, min = 0, max = 30), ncol = 3)
getProportion(table = myTable)
}

