% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSatellitalMaps.R
\name{getSatellitalMaps}
\alias{getSatellitalMaps}
\title{Obtener información satelital de variables oceanográficas}
\usage{
getSatellitalMaps(initialDate = NULL, finalDate = NULL, timeRes = "month",
  what = "SST", dateList = NULL, lonRange = c(-85, -70),
  latRange = c(-20, -2), outputFormat = "png", outputDir = ".")
}
\arguments{
\item{initialDate}{Fecha inicial para la extracción de mapas.}

\item{finalDate}{Fecha final para la extracción de mapas.}

\item{timeRes}{Resolución temporal para la obtención de mapas: Mensual ('month') o diario ('day').}

\item{what}{¿Qué variable desea descargar? Ver detalles.}

\item{dateList}{Lista de fechas para la extracción de mapas.}

\item{lonRange}{Rango de coordenadas de longitud. Si es \code{NULL}, tomará valores entre 85°W y 90°W.}

\item{latRange}{Rango de coordenadas de latitud. Si es \code{NULL}, tomará valores entre 20°S y 2°S.}

\item{outputFormat}{Formato de descarga de archivos: NetCDF (nc), Valores separados por comas (csv),
image (png).}

\item{outputDir}{Directorio para la descarga de archivos.}
}
\description{
Esta función toma un rango de fechas y coordenadas y extrae los mapas de distribución de
Temperatura Superficial del Mar (SST), Salinidad Superficial del Mar (SSS), Clorofila-a (chl) y
Topografía (topo)
}
\examples{
getSatellitalMaps(initialDate = "2014-3-26", finalDate = "2014-4-17", what = "sst", timeRes = "day",
lonRange = c(-90, -85), latRange = c(-5, 0), outputFormat = "png",
outputDir = "D:/output/")
}

