% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSatellitalMaps.R
\name{getSatellitalMaps}
\alias{getSatellitalMaps}
\title{Get satellital information for oceanographic variables}
\usage{
getSatellitalMaps(initialDate = NULL, finalDate = NULL, timeRes = "month",
  what = "SST", atDepth = NULL, dateList = NULL, lonRange = c(-85, -70),
  latRange = c(-20, -2), outputFormat = "png", outputDir = ".",
  showURL = FALSE)
}
\arguments{
\item{initialDate}{Initial date Fecha inicial para la extracción de mapas.}

\item{finalDate}{Fecha final para la extracción de mapas.}

\item{timeRes}{Resolución temporal para la obtención de mapas: Mensual ('month') o diario ('day').}

\item{what}{¿Qué variable desea descargar? Ver detalles.}

\item{atDepth}{Define depth value for getting info.}

\item{dateList}{Lista de fechas para la extracción de mapas.}

\item{lonRange}{Rango de coordenadas de longitud. Si es \code{NULL}, tomará valores entre 85°W y 90°W.}

\item{latRange}{Rango de coordenadas de latitud. Si es \code{NULL}, tomará valores entre 20°S y 2°S.}

\item{outputFormat}{Formato de descarga de archivos: NetCDF (nc), Valores separados por comas (csv),
image (png).}

\item{outputDir}{Directorio para la descarga de archivos.}

\item{showURL}{Do you want to show the URL each time?}
}
\description{
This function uses date and space range and get maps for Sea Surface Temperature (SST),
Sea Surface Salinity (SSS), Chlorophyll-a (Chl), Topography (topo), Sea level (sealevel) and SODA
estimations for Temperature (sst_soda) and Salinity (sss_soda) by depth (\code{atDepth} parameter).
}
\examples{
getSatellitalMaps(initialDate = "2014-3-26", finalDate = "2014-4-17", what = "sst", timeRes = "day",
lonRange = c(-90, -85), latRange = c(-5, 0), outputFormat = "png",
outputDir = "../")
}

