% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isopAreaAssigner.R
\name{isopArea.assigner}
\alias{isopArea.assigner}
\title{Title Function to get AIP code from lon-lat information.}
\usage{
isopArea.assigner(dataPoints, colLon = "lon", colLat = "lat", old = TRUE)
}
\arguments{
\item{dataPoints}{\code{data.frame} which has Longitude and Latitude information.}

\item{colLon}{Name of column which contains Longitude info.}

\item{colLat}{Name of column which contains Latitude info.}

\item{old}{\code{logical}. ¿Desea utilizar la base de datos antigua (\code{old = TRUE}) o nueva \code{old = FALSE}?}
}
\description{
Title Function to get AIP code from lon-lat information.
}

