% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthFreq.R
\name{lengthFrequencyPlot}
\alias{lengthFrequencyPlot}
\title{Given a frequency table size, it generates a vertical array of graphics for each specified
category (months, latitudes, years, etc.).}
\usage{
lengthFrequencyPlot(file1, file2 = NULL, profile = NULL, xlim = NULL,
  xInterval = 1, ylim = c(0, 50), yInterval = NULL, ltys1 = "solid",
  lwds1 = "1", cols1 = "black", ltys2 = "solid", lwds2 = "1",
  cols2 = "blue", juvLine = NULL, juvLty = "dotted", juvLwd = 1,
  juvCol = "red", cex.axis = 1, cex.lab = 1, smooth = FALSE,
  oma = c(5, 5, 1, 3), xLab = "Longitud total (cm)",
  yLab = "Frecuencia (\%)", noDataLabel = "Sin datos",
  juvLabel = "juveniles = ", yLabFactor = 1, relative = TRUE)
}
\arguments{
\item{file1}{Indica la tabla desde donde se dibujarán las frecuencias por talla. Ver detalles.}

\item{file2}{Indica la tabla secundaria desde donde se dibujarán las frecuencias por talla. Ver detalles.}

\item{xlab}{}
}
\details{
This function takes a length frequency data and returns an stacking plot of that frequencies.
smooth: logical. If TRUE, function uses an spline function to plot by categories.
This function uses a data with categories at cols and length at rows.
}

