% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LengthFreq.R
\name{lengthFrequencyPlot}
\alias{lengthFrequencyPlot}
\title{Given a frequency table size, it generates a vertical array of graphics for each specified
category (months, latitudes, years, etc.).}
\usage{
lengthFrequencyPlot(file, file2 = NA, profile = NULL,
  categoryNames = NULL, xlab = "Longitud (cm)", ylab = "Frecuencia (\%)",
  ylim = NULL, cex.axis = 1.2, xlim = NULL, Xinterval = 1,
  Yinterval = 10, lcol = "blue", lcol2 = NULL, lwd = 1, lty = "solid",
  lineCategory = -2, styleLab = 2, jValue = NA, jtext = NA, jcex = 1,
  jtextcol = "black", cat.cex = 1, jround = 1, jcol = "red", jwd = 1,
  jty = "dotted", sinpesca = "Sin pesca", sinjuveniles = "Sin juveniles",
  zeros = FALSE, nzeros = 0, showRange = TRUE, adj = 0.02,
  showJuv = TRUE, smooth = TRUE, spar = 0.01, totalFreq = FALSE,
  totalLabel = "Total")
}
\arguments{
\item{file}{\code{character}. File path of length frequencies table.}

\item{file2}{\code{character}. File path of second length frequencies table.}

\item{profile}{\code{character}. Profile by species.}

\item{xlab}{A title for the X axis.}

\item{ylab}{A title for the Y axis.}

\item{cex.axis}{Size of axis text.}

\item{xlim}{Limits for the X axis.}

\item{Xinterval}{Intervals for the X axis.}

\item{Yinterval}{Intervals for the Y axis.}

\item{lcol}{Vector of colors for length lines (one or more).}

\item{lcol2}{Color (just one) of length line for second table.}

\item{lwd}{Width of length lines.}

\item{lty}{Vector of line type for length lines (one or more).}

\item{styleLab}{Numer (0,1,2,3), the style of axis labels.}

\item{jValue}{Juvenile length.}

\item{totalLabel}{}
}
\details{
This function takes a length frequency data and returns an stacking plot of that frequencies.
smooth: logical. If TRUE, function uses an spline function to plot by categories.
This function uses a data with categories at cols and length at rows.
}

