% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{lengthFrequencyPlot}
\alias{lengthFrequencyPlot}
\title{Given a frequency table size, it generates a vertical array of graphics for each specified
category (months, latitudes, years, etc.).}
\usage{
lengthFrequencyPlot(file1, file2 = NULL, profile = NULL, xlim = NULL,
  xInterval = 1, ylim = c(0, 50), yInterval = NULL, ltys1 = "solid",
  lwds1 = "1", cols1 = "black", ltys2 = "solid", lwds2 = "1",
  cols2 = "blue", juvLine = NULL, juvLty = "dotted", juvLwd = 1,
  juvCol = "red", juvCex = 1, yLab_line = 3, cex.axis = 1,
  cex.lab = 1, juvAdj = 0.99, juvRound = 0, namesAdj = 0.01,
  smooth = FALSE, oma = c(5, 5, 1, 3), xLab = NULL,
  yLab = "Frecuencia (\%)", noDataLabel = "Sin datos",
  juvLabel = "juveniles = ", yLabFactor = 1, relative = TRUE)
}
\arguments{
\item{file1}{Indicates the table from the main length distribution will be drawn. See details.}

\item{file2}{Indicates the table from the secondary length distribution will be drawn. See details.}

\item{profile}{\code{character}. Indicates the profile (as an species' name) used to predine values of
xlim, juvLine, xInterval and xLab.}

\item{xlim}{Limits of x axis.}

\item{xInterval}{Number of intervals for x axis labels.}

\item{ylim}{Limits of y axis.}

\item{yInterval}{Number of intervals for y axis labels.}

\item{ltys1}{lty (see \code{\link{par}}) parameter for main length distribution lines.}

\item{lwds1}{lwd (see \code{\link{par}}) parameter for main length distribution lines.}

\item{cols1}{col (see \code{\link{par}}) parameter for main length distribution lines.}

\item{ltys2}{lty (see \code{\link{par}}) parameter for secondary length distribution lines.}

\item{lwds2}{lwd (see \code{\link{par}}) parameter for secondary length distribution lines.}

\item{cols2}{col (see \code{\link{par}}) parameter for secondary length distribution lines.}

\item{juvLine}{line (see \code{\link{mtext}}) parameter for juvenile text.}

\item{juvLty}{lty (see \code{\link{par}}) parameter for juvenile line.}

\item{juvLwd}{lwd (see \code{\link{par}}) parameter for juvenile line.}

\item{juvCol}{col (see \code{\link{par}}) parameter for juvenile line.}

\item{juvCex}{Size of juveniles' text.}

\item{yLab_line}{line (see \code{\link{mtext}}) parameter for ylab text.}

\item{cex.axis}{cex.axis (see \code{\link{par}}) parameter for x and y axes.}

\item{cex.lab}{cex.lab (see \code{\link{par}}) parameter for x and y axes.}

\item{juvAdj}{adj (see \code{\link{mtext}}) parameter for juvenile text.}

\item{juvRound}{Number of decimals places used to show juveniles' values.}

\item{namesAdj}{adj (see \code{\link{mtext}}) parameter for categories (column names) text.}

\item{smooth}{\code{logical} Do you want to smooth the length distribution lines?}

\item{oma}{oma (see \code{\link{par}}) parameter for plot.}

\item{xLab}{Label for x axis.}

\item{yLab}{Label for y axis.}

\item{noDataLabel}{\code{character}. If there is no data in a column, the function will show
what this parameter indicates.}

\item{juvLabel}{Text before juvenile value.}

\item{yLabFactor}{\code{numeric}. Value that multiplies all values in both main and secondary matrices.}

\item{relative}{\code{logical} Do you want to plot relatives or absolutes values?}
}
\details{
This function takes a length frequency data and returns an stacking plot of that frequencies.
smooth: logical. If TRUE, function uses an spline function to plot by categories.
This function uses a data with categories at cols and length at rows.
}

