% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{minDistanceToCoast}
\alias{minDistanceToCoast}
\title{Get the minumim distance to coast}
\usage{
minDistanceToCoast(data, colLon = "lon", colLat = "lat",
  countryFilter = "peru", unit = "nm")
}
\arguments{
\item{data}{\code{data.frame} with coords that will be used to calculate min distance to coast line.}

\item{colLon}{Name or position of column for longitude. As default, it will be \code{lon}.}

\item{colLat}{Name or position of column for latitude. As default, it will be \code{lat}.}

\item{countryFilter}{Select the country for make comparation}

\item{unit}{Define the unit for outputs: nm (nautical miles), kilometers (km), m (meters).}
}
\value{
It returns a list with both the minimum distance to coast line and the point where this
distance is reached.
}
\examples{

n <- 100

allData <- data.frame(lon = runif(n, -80, -70), lat = runif(n, -18, -2),
                      stringsAsFactors = FALSE)

allData <- allData[!is.na(isopArea.assigner(allData)),]

minValues <- minDistanceToCoast(allData)

xlim <- c(-85, -70)
ylim <- c(-20, -2)

dev.new()
par(mar = c(2, 3, 1, 1), xaxs = "i", yaxs = "i")
plot(1, 1, pch = NA, axes = FALSE, xlab = NA, ylab = NA, xlim = xlim, ylim = ylim)

points(allData$lon, allData$lat, pch = 16, cex = 0.5, xlim = c(-85, -70), ylim = c(-20, -2))
lines(coastline$lon, coastline$lat)


for(i in 1:nrow(minValues$position)){
  lines(c(allData$lon[i], minValues$position$lon[i]), c(allData$lat[i], minValues$position$lat[i]),
        lty = "dotted", col = "red")
}

axis(side = 1, at = seq(xlim[1], xlim[2], length.out = 4),
     labels = getCoordsAxes(seq(xlim[1], xlim[2], length.out = 4), "lon"))
axis(side = 2, at = seq(ylim[1], ylim[2], length.out = 10),
     labels = getCoordsAxes(seq(ylim[1], ylim[2], length.out = 10), "lat"), las = 2)
box()
}
