% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{minDistanceToCoast}
\alias{minDistanceToCoast}
\title{Get the minumim distance to coast}
\usage{
minDistanceToCoast(data, colLon = "lon", colLat = "lat",
  countryFilter = "peru", unit = "nm", multicore = FALSE,
  ncores = 1, out = c("value", "position"))
}
\arguments{
\item{data}{\code{data.frame} with coords that will be used to calculate min distance to coast line.}

\item{colLon}{Name or position of column for longitude. As default, it will be \code{lon}.}

\item{colLat}{Name or position of column for latitude. As default, it will be \code{lat}.}

\item{countryFilter}{Select the country for make comparation}

\item{unit}{Define the unit for outputs: nm (nautical miles), kilometers (km), m (meters).}

\item{multicore}{\code{logical} indicating whether to use multiple cores for calculate distances.
See Details.}

\item{ncores}{If \code{multicore = TRUE}, how many cores are you going to use?}
}
\value{
It returns a list with 2 levels: \code{value}, The minimum distance to coast line and
\code{position}, the point where this minimum distance is reached.
}
\description{
Get the minumim distance to coast
}
\details{
This function uses internaly both \link{spDists} and \link{spDistsN1}, with argument
\code{longlat = TRUE}.

If \code{data} is a matrix, \code{colLon} and \code{colLat} must be ONLY numeric values.

It is recommended to use multicore mode only when the database exceeds 2000 rows, less than
this, single core mode have proved to be faster.
}
\examples{

n <- 100

allData <- data.frame(lon = runif(n, -80, -70), lat = runif(n, -18, -2),
                      stringsAsFactors = FALSE)

allData <- allData[!is.na(isopArea.assigner(allData)),]

minValues <- minDistanceToCoast(allData)

xlim <- c(-85, -70)
ylim <- c(-20, -2)

dev.new()
par(mar = c(2, 3, 1, 1), xaxs = "i", yaxs = "i")
plot(1, 1, pch = NA, axes = FALSE, xlab = NA, ylab = NA, xlim = xlim, ylim = ylim)

points(allData$lon, allData$lat, pch = 16, cex = 0.5, xlim = c(-85, -70), ylim = c(-20, -2))
lines(coastline$lon, coastline$lat)


for(i in 1:nrow(minValues$position)){
  lines(c(allData$lon[i], minValues$position$lon[i]), c(allData$lat[i], minValues$position$lat[i]),
        lty = "dotted", col = "red")
}

axis(side = 1, at = seq(xlim[1], xlim[2], length.out = 4),
     labels = getCoordsAxes(seq(xlim[1], xlim[2], length.out = 4), "lon"))
axis(side = 2, at = seq(ylim[1], ylim[2], length.out = 10),
     labels = getCoordsAxes(seq(ylim[1], ylim[2], length.out = 10), "lat"), las = 2)
box()
}
