% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{progressBar}
\alias{progressBar}
\title{Make a progress bar}
\usage{
progressBar(i, n, stepText = "n")
}
\arguments{
\item{i}{Step of the iteration.}

\item{n}{Total number of iteration}
}
\value{
The function returns only messages (from \code{cat}).
}
\description{
Make a progress bar
}
\examples{
n = 1000
for(i in seq(n)){
  Sys.sleep(0.01)
  progressBar(i = i, n = n)
}
}

