% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRectangles.R
\name{randomRectangles}
\alias{randomRectangles}
\title{Draw colorful squares}
\usage{
randomRectangles(nsquares = 500, borders = TRUE,
  colPalette = rainbow(n = 2000))
}
\arguments{
\item{nsquares}{\code{numeric} How much squares do you desire? (500, as default).}

\item{borders}{\code{logical} Do you want yo show borders around rectangles?}

\item{colPalette}{\code{character} Vector of colors that will be used for plotting.}
}
\description{
Draw colorful squares
}
\examples{
randomRectangles(nsquares = 120)
}
