% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{package}
\name{ruisu-package}
\alias{ruisu-package}
\alias{ruisu}
\title{Miscellany functions for the IMARPE work}
\description{
Package with useful tools and functions for everyday IMARPE work.
}
\author{
Wencheng Lau-Medrano, \email{llau@imarpe.gob.pe}
}
\keyword{IMARPE}
\keyword{miscellany,}
