% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{package}
\name{ruisu-package}
\alias{ruisu}
\alias{ruisu-package}
\title{Paquete con herramientas diversas.}
\description{
Paquete con funciones misceláneas dentro del Instituto del Mar del Perú.
}
\author{
Wencheng Lau-Medrano, \email{llau@imarpe.gob.pe}
}
\keyword{IMARPE}
\keyword{miscelánea,}

