% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\docType{package}
\name{ruisu-package}
\alias{ruisu-package}
\alias{ruisu}
\title{Tools for habitual tasks (and another fancy things) at IMARPE}
\description{
Tools for common tasks as well as some fancy functions
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@gmail.com}
}
\keyword{IMARPE}
\keyword{miscellany,}
