% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPtsByPolygon.R
\name{selectPtsByPolygon}
\alias{selectPtsByPolygon}
\title{Function to obtain an index of points inside and outside a polygon}
\usage{
selectPtsByPolygon(x, n_polygon, coords_x = c("x", "y"), crs = 4326, ...)
}
\arguments{
\item{x}{\code{data.frame} with coordinates to be classified.}

\item{n_polygon}{Number of points to define polygon.}

\item{coords_x}{Names of coordinates on \code{x}.}

\item{crs}{Coordinate reference system}

\item{...}{Extra arguments passed to the main plot.}
}
\value{
A numeric vector: out of (\code{0}) or within (\code{1}) the polygon.
}
\description{
Function to obtain an index of points inside and outside a polygon
}
\examples{
n <- 50
sampleData <- data.frame(x = runif(n = 50, min = 0.5, max = 1.5),
                         y = runif(n = 50, min = 0.5, max = 1.5))

index <- selectPtsByPolygon(x = sampleData, n_polygon = 8, asp = 1)

plot(x = sampleData$x, y = sampleData$y,
     col = ifelse(index == 0, "black", "blue"), asp = 1)
}
