% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ruisu.R
\name{sillyTest}
\alias{sillyTest}
\title{Functions for making an interactive test}
\usage{
sillyTest(testName, questions, answers, correctAnswers)
}
\arguments{
\item{testName}{The name for the test}

\item{questions}{\code{character} vector indicating the questions.}

\item{answers}{\code{list} with the options for each question.}

\item{correctAnswers}{\code{numeric} vector indicating the index for correct answers.}
}
\value{
Nothing.
}
\description{
Functions for making an interactive test
}
\details{
Of course, arguments for questions, answers and correct answers must have the same length.
}
\examples{
sillyTest(testName = "Harry Potter", questions = testExample$questions,
          answers = testExample$answers, correctAnswers = testExample$correctAnswers)
}
